bytes: consolidate streambuf bytes utilities
The input/output variants have the exact same header dependencies.
