+++ src/v/bytes/tests/iobuf_tests.cc
@@ -10,11 +10,10 @@
 #include "bytes/bytes.h"
 #include "bytes/details/io_allocation_size.h"
 #include "bytes/iobuf.h"
-#include "bytes/iobuf_istreambuf.h"
-#include "bytes/iobuf_ostreambuf.h"
 #include "bytes/iobuf_parser.h"
 #include "bytes/iostream.h"
 #include "bytes/scattered_message.h"
+#include "bytes/streambuf.h"
 #include "bytes/tests/utils.h"
 #include "bytes/utils.h"
