+++ src/v/cloud_storage/partition_manifest.cc
@@ -11,9 +11,8 @@
 #include "cloud_storage/partition_manifest.h"
 
 #include "bytes/iobuf.h"
-#include "bytes/iobuf_istreambuf.h"
-#include "bytes/iobuf_ostreambuf.h"
 #include "bytes/iostream.h"
+#include "bytes/streambuf.h"
 #include "cloud_storage/base_manifest.h"
 #include "cloud_storage/logger.h"
 #include "cloud_storage/segment_meta_cstore.h"
